Links added on main page and Confirm box before delete msg  and Hide comment on click edit button

+++ resources/views/main/my_page/my_info/novel_group_comments_manage.blade.php
@@ -34,39 +34,45 @@
                         <!-- //댓글정렬 -->
                     </div>
                     <ul class="comment-list">
-                        @foreach($novel_comments as $novel_comment)
-                            <li>
-                                <div class="comment-wrap">
-                                    <div class="comment-info"><span
-                                                class="parent-subject">{{$novel_comment->inning}} {{$novel_comment->novel_title}}</span><span
-                                                class="writer">{{$novel_comment->user_name}}</span></div>
-                                    <div class="comment-btns"><a href="#mode_nav"
-                                                                 v-on:click="comment_box_show({{$novel_comment->id}})">수정</a><a
-                                                href="#mode_nav" v-on:click="remove_comment('{{$novel_comment->id}}')">삭제</a>
-                                    </div>
-                                    <div class="comment-content">
-                                        <p>{{$novel_comment->comment}}</p>
-                                    </div>
-                                    <div class="comment-content " id="comment_box{{$novel_comment->id}}"
-                                         v-if="display.id =={{$novel_comment->id}} && display.status">
+                        @if(count($novel_comments) > 0)
+                            @foreach($novel_comments as $novel_comment)
+                                <li>
+                                    <div class="comment-wrap">
+                                        <div class="comment-info"><span
+                                                    class="parent-subject">{{$novel_comment->inning}} {{$novel_comment->novel_title}}</span><span
+                                                    class="writer">{{$novel_comment->user_name}}</span></div>
+                                        <div class="comment-btns"><a href="#mode_nav"
+                                                                     v-on:click="comment_box_show({{$novel_comment->id}})">수정</a><a
+                                                    href="#mode_nav"
+                                                    v-on:click="remove_comment('{{$novel_comment->id}}')">삭제</a>
+                                        </div>
+                                        <div class="comment-content" v-show="!display.status">
+                                            <p>{{$novel_comment->comment}}</p>
+                                        </div>
+                                        <div class="comment-content " id="comment_box{{$novel_comment->id}}"
+                                             v-if="display.id =={{$novel_comment->id}} && display.status">
 
                                          <textarea name="comment" id="comment{{$novel_comment->id}}" rows="3"
                                                    style="width:65%;">{{$novel_comment->comment}}</textarea>
 
-                                        <button name="edit" id="edit{{$novel_comment->id}}"
-                                                v-on:click="update_comment('{{$novel_comment->id}}')"
-                                                class="btn btn-primary inline"
-                                                style="width:100px;height:51px;vertical-align: top;">
-                                            Edit
-                                        </button>
-                                    </div>
+                                            <button name="edit" id="edit{{$novel_comment->id}}"
+                                                    v-on:click="update_comment('{{$novel_comment->id}}')"
+                                                    class="btn btn-primary inline"
+                                                    style="width:100px;height:51px;vertical-align: top;">
+                                                Edit
+                                            </button>
+                                        </div>
 
-                                    <div class="comment-etc-info">{{str_limit($novel_comment->novel_group_title,70)}}<span
-                                                class="datetime">{{time_elapsed_string($novel_comment->created_at)}}</span>
+                                        <div class="comment-etc-info">{{str_limit($novel_comment->novel_group_title,70)}}
+                                            <span
+                                                    class="datetime">{{time_elapsed_string($novel_comment->created_at)}}</span>
+                                        </div>
                                     </div>
-                                </div>
-                            </li>
-                        @endforeach
+                                </li>
+                            @endforeach
+                        @else
+                            <div class="comment-wrap" style="text-align: center"> 첫 번째 댓글을 작성해 보세요.</div>
+                        @endif
 
                     </ul>
                 </div>
@@ -106,12 +112,16 @@
                 }
             },
             remove_comment: function (comment_id) {
-                app.$http.delete('{{ url('comments') }}/' + comment_id, {headers: {'X-CSRF-TOKEN': window.Laravel.csrfToken}})
-                        .then(function (response) {
-                            location.reload();
-                        }).catch(function (errors) {
-                            console.log(errors);
-                        });
+
+                if (confirm('Are you sure to delete this comment?')) {
+
+                    app.$http.delete('{{ url('comments') }}/' + comment_id, {headers: {'X-CSRF-TOKEN': window.Laravel.csrfToken}})
+                            .then(function (response) {
+                                location.reload();
+                            }).catch(function (errors) {
+                                console.log(errors);
+                            });
+                }
             },
 
             update_comment: function (comment_id, comment_type) {
@@ -119,7 +129,7 @@
                 this.info.comment_type = comment_type;
                 this.info.comment = $('#comment' + comment_id).val();
 
-                app.$http.put('{{ url('comments') }}/'+comment_id,  this.info, {headers: {'X-CSRF-TOKEN': window.Laravel.csrfToken}})
+                app.$http.put('{{ url('comments') }}/' + comment_id, this.info, {headers: {'X-CSRF-TOKEN': window.Laravel.csrfToken}})
                         .then(function (response) {
                             location.reload();
